backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
crewai[google-genai]
google-genai
